Fixed error that was causing non auth request to go to login in case of 401 error

+++ Trivia Patente/Trivia Patente/HTTPManager.swift
@@ -45,7 +45,7 @@ class HTTPManager {
                             handler(response)
                         } else if let message = (response.result.error as? BackendError)?.message {
                             let response = T(error: message, statusCode: response.response?.statusCode)
-                            if response.statusCode == 401 {
+                            if response.statusCode == 401 && auth {
                                 SessionManager.drop()
                                 UIViewController.goToFirstAccess()
                             } else {
